+++ tests/integration/berd/test_berd_short.py
@@ -3,7 +3,7 @@ import unittest
 
 from app.definitions.data import SurveyMetadata, SPP, ImageResponse
 from app.controllers.flat import get_pck
-from tests.integration.mapped import read_submission_data
+from tests.integration.flat import read_submission_data
 
 
 class BerdTests(unittest.TestCase):
