+++ tests/integration/flat/test_blocks.py
@@ -2,27 +2,24 @@ import unittest
 
 from app.definitions.data import SurveyMetadata, PCK
 from app.controllers.flat import get_pck
-from tests.integration.mapped import read_submission_data, are_equal
+from tests.integration.flat import read_submission_data, are_equal
 
 
-class QPSESRAPPckTests(unittest.TestCase):
-
-    def test_0003_to_pck(self):
-        filepath = "tests/data/qpsesrap/169.0003.json"
+class BlocksPckTests(unittest.TestCase):
+    def test_0001_to_pck(self):
+        filepath = "tests/data/blocks/073.0002.json"
         submission_data = read_submission_data(filepath)
 
         survey_metadata: SurveyMetadata = {
-            "survey_id": "169",
-            "period_id": "1605",
+            "survey_id": "073",
+            "period_id": "201605",
             "ru_ref": "12346789012A",
-            "form_type": "0003",
+            "form_type": "0002",
             "period_start_date": "2016-05-01",
             "period_end_date": "2016-05-31",
         }
-
         actual: PCK = get_pck(submission_data, survey_metadata)
-
-        pck_filepath = "tests/data/qpsesrap/169.0003.pck"
+        pck_filepath = "tests/data/blocks/073.0002.pck"
         with open(pck_filepath) as f:
             expected: PCK = f.read()
