+++ tests/integration/flat/test_mbs_spp.py
@@ -3,7 +3,7 @@ import unittest
 
 from app.definitions.data import SurveyMetadata, PCK
 from app.controllers.flat import get_pck
-from tests.integration.mapped import read_submission_data
+from tests.integration.flat import read_submission_data
 
 
 class MbsSppTests(unittest.TestCase):
