+++ tests/integration/flat/test_qpsespb.py
@@ -2,17 +2,17 @@ import unittest
 
 from app.definitions.data import SurveyMetadata, PCK
 from app.controllers.flat import get_pck
-from tests.integration.mapped import read_submission_data, are_equal
+from tests.integration.flat import read_submission_data, are_equal
 
 
-class QPSESPckTests(unittest.TestCase):
+class QPSESPBPckTests(unittest.TestCase):
 
     def test_0002_to_pck(self):
-        filepath = "tests/data/qpses/160.0002.json"
+        filepath = "tests/data/qpsespb/165.0002.json"
         submission_data = read_submission_data(filepath)
 
         survey_metadata: SurveyMetadata = {
-            "survey_id": "160",
+            "survey_id": "165",
             "period_id": "1605",
             "ru_ref": "12346789012A",
             "form_type": "0002",
@@ -22,7 +22,7 @@ class QPSESPckTests(unittest.TestCase):
 
         actual: PCK = get_pck(submission_data, survey_metadata)
 
-        pck_filepath = "tests/data/qpses/160.0002.pck"
+        pck_filepath = "tests/data/qpsespb/165.0002.pck"
         with open(pck_filepath) as f:
             expected: PCK = f.read()
