+++ tests/integration/flat/test_rails.py
@@ -3,20 +3,20 @@ import unittest
 
 from app.definitions.data import SurveyMetadata
 from app.controllers.flat import get_pck
-from tests.integration.mapped import read_submission_data
+from tests.integration.flat import read_submission_data
 
 
-class FuelsTests(unittest.TestCase):
+class RailsTests(unittest.TestCase):
 
-    def test_0002_to_json(self):
-        filepath = "tests/data/fuels/024.0002.json"
+    def test_0001_to_json(self):
+        filepath = "tests/data/rails/194.0001.json"
         submission_data = read_submission_data(filepath)
 
         survey_metadata: SurveyMetadata = {
-            "survey_id": "024",
+            "survey_id": "194",
             "period_id": "201605",
             "ru_ref": "12346789012A",
-            "form_type": "0002",
+            "form_type": "0001",
             "period_start_date": "2016-05-01",
             "period_end_date": "2016-05-31",
         }
@@ -24,7 +24,7 @@ class FuelsTests(unittest.TestCase):
         result: str = get_pck(submission_data, survey_metadata)
         actual: dict[str, str] = json.loads(result)
 
-        pck_filepath = "tests/data/fuels/result.json"
+        pck_filepath = "tests/data/rails/result.json"
         with open(pck_filepath) as f:
             expected: dict[str, str] = json.load(f)
